Fixed functionality. Added tests, and updated properties to be longs

+++ src/ZendeskApi.Client/Models/AuditEvents/AttachmentRedactionEvent.cs
@@ -5,8 +5,8 @@ namespace ZendeskApi.Client.Models.AuditEvents
     public class AttachmentRedactionEvent : AuditEvent, IAuditEvent
     {
         [JsonProperty("attachment_id")]
-        public int AttachmentId { get; set; }
+        public long AttachmentId { get; set; }
         [JsonProperty("comment_id")]
-        public int CommentId { get; set; }
+        public long CommentId { get; set; }
     }
 }
