+++ src/ZendeskApi.Client/Models/AuditEvents/CCEvent.cs
@@ -7,7 +7,7 @@ namespace ZendeskApi.Client.Models.AuditEvents
         [JsonProperty("body")]
         public string Body { get; set; }
         [JsonProperty("recipients")]
-        public int[] Recipients { get; set; }
+        public long[] Recipients { get; set; }
         [JsonProperty("via")]
         public Via Via { get; set; }
     }
