+++ src/ZendeskApi.Client/Models/AuditEvents/IAuditEvent.cs
@@ -1,12 +1,13 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
+using ZendeskApi.Client.Converters;
 
 namespace ZendeskApi.Client.Models.AuditEvents
 {
     public interface IAuditEvent
     {
         [JsonProperty("id")]
-        int Id { get; set; }
+        long Id { get; set; }
         [JsonProperty("type")]
         AuditTypes Type { get; set; }
     }
